test: all upload-button stories e2e sanity - done
