+++ cypress/support/iframe.js
@@ -4,7 +4,15 @@ const command = (selector) =>
     cy.get(selector, { log: false })
         .its("0.contentDocument.body", { log: false })
         .should("not.be.empty")
-        .then(cy.wrap);
+        .then((iframeBody)=>{
+
+            cy.window().then((appWindow) => {
+                iframeBody.ownerDocument.defaultView
+                    .XMLHttpRequest = appWindow.XMLHttpRequest
+            });
+
+            return cy.wrap(iframeBody);
+        });
 
 Cypress.Commands.add("iframe",command);
