+++ package.json
@@ -26,6 +26,7 @@
     "test:ci": "yarn flow && yarn lint && yarn jest:ci",
     "cy:open": "CYPRESS_SB_PORT=9111 cypress open",
     "cy:run": "wait-on http://localhost:9111 -d 3000 && CYPRESS_SB_PORT=9111 cypress run",
+    "cy:run:ff": "yarn cy:run --browser firefox",
     "e2e": "concurrently --success first --kill-others \"yarn sb:start\" \"yarn cy:run\"",
     "server": "cd server; PORT=4000 node index.js"
   },
